feat: aws audit logs solution
