+++ examples/simple/02-test.tf
@@ -1,96 +1,95 @@
-resource "azurerm_service_plan" "this" {
-  name                     = "${local.project}-asp"
-  location                 = var.location
-  resource_group_name      = azurerm_resource_group.rg.name
-  sku_name                 = "P1v3"
-  worker_count             = 1
-  os_type                  = "Linux"
-  per_site_scaling_enabled = false
-  zone_balancing_enabled   = false
-  tags                     = var.tags
+data "archive_file" "this" {
+  type        = "zip"
+  source_file = "./LogGenerator/index.py"
+  output_path = "./LogGenerator/lambda.zip"
 }
 
-resource "azurerm_storage_account" "this" {
-  name                     = replace("${local.project}funcst", "-", "")
-  location                 = var.location
-  resource_group_name      = azurerm_resource_group.rg.name
-  account_tier             = "Standard"
-  account_replication_type = "LRS"
-  min_tls_version          = "TLS1_2"
+# resource "aws_cloudwatch_log_group" "this" {
+#   name = "${local.project}-log-group"
+  
+#   retention_in_days = 14
+
+#   tags = {
+#     Name = "auditlogs"
+#   }
+# }
+
+# resource "aws_cloudwatch_log_stream" "this" {
+#   name           = "${local.project}-log-stream"
+#   log_group_name = aws_cloudwatch_log_group.this.name
+# }
+
+resource "aws_cloudwatch_event_rule" "this" {
+  name        = "${local.project}-audit-rule"
+
+  schedule_expression = "rate(1 minute)"
 }
 
-resource "azurerm_linux_function_app" "this" {
-  name                = "${local.project}-func"
-  resource_group_name = azurerm_resource_group.rg.name
-  location            = var.location
+resource "aws_cloudwatch_event_target" "this" {
+  rule      = aws_cloudwatch_event_rule.this.name
+  target_id = "TriggerLambda"
+  arn       = aws_lambda_function.lambda_function.arn
+}
 
-  storage_account_name          = azurerm_storage_account.this.name
-  storage_uses_managed_identity = true
-  service_plan_id               = azurerm_service_plan.this.id
+resource "aws_iam_role" "lambda" {
+  name = var.lambda.role_name
+ 
+  assume_role_policy = jsonencode({
+    Version = "2012-10-17",
+    Statement = [{
+      Effect = "Allow",
+      Principal = {
+        Service = "lambda.amazonaws.com"
+      },
+      Action = ["sts:AssumeRole"]
+    }]
+  })
+}
 
-  client_certificate_enabled = false
-  https_only                 = true
-
-  site_config {
-    always_on                              = true
-    use_32_bit_worker                      = false
-    ftps_state                             = "Disabled"
-    http2_enabled                          = true
-    minimum_tls_version                    = "1.2"
-    scm_minimum_tls_version                = "1.2"
-    vnet_route_all_enabled                 = true
-    application_insights_connection_string = azurerm_application_insights.this.connection_string
-    application_insights_key               = azurerm_application_insights.this.instrumentation_key
-
-    application_stack {
-      node_version = "20"
-    }
-
-    cors {
-      allowed_origins = [
-        "https://portal.azure.com",
-      ]
-      support_credentials = false
-    }
-  }
-
-  app_settings = {
-    APPINSIGHTS_SAMPLING_PERCENTAGE = 5
-    WEBSITE_RUN_FROM_PACKAGE        = 1
-  }
-
-  identity {
-    type = "SystemAssigned"
-  }
-
-  lifecycle {
-    ignore_changes = [
-      tags["hidden-link: /app-insights-conn-string"],
-      tags["hidden-link: /app-insights-instrumentation-key"],
-      tags["hidden-link: /app-insights-resource-id"],
+resource "aws_iam_policy" "lambda" {
+  name = var.lambda.policy_name
+ 
+  policy = jsonencode({
+    Version = "2012-10-17",
+    Statement = [
+      {
+        Action = [
+          "logs:CreateLogStream",
+          "logs:CreateLogGroup",
+          "logs:PutLogEvents"
+        ],
+        Effect   = "Allow",
+        Resource = "arn:aws:logs:*:*:*"
+      }
     ]
+  })
+}
+
+resource "aws_iam_role_policy_attachment" "function_logging_policy_attachment" {
+  role       = aws_iam_role.lambda.id
+  policy_arn = aws_iam_policy.lambda.arn
+}
+
+resource "aws_lambda_function" "lambda_function" {
+  filename      = "./lambda.zip"
+  function_name = "test_lambda_logs"
+  role          = aws_iam_role.lambda.arn
+  handler       = "index.lambda_handler"
+  runtime       = "python3.12"
+  timeout       = 900
+  environment {
+    variables = {
+      log_group_name  = "${local.log_group_name}"
+      log_stream_name = "${local.log_stream_name}"
+      log_throuput    =  "${local.log_throuput}"
+    }
   }
 }
 
-resource "azurerm_role_assignment" "this" {
-  scope                = azurerm_storage_account.this.id
-  role_definition_name = "Storage Blob Data Contributor"
-  principal_id         = azurerm_linux_function_app.this.identity[0].principal_id
-}
-
-resource "null_resource" "deploy" {
-  depends_on = [azurerm_role_assignment.this]
-
-  triggers = {
-    deploy_version = "1.1" # change me to redeploy
-    function_name  = azurerm_linux_function_app.this.name
-  }
-
-  provisioner "local-exec" {
-    command = <<EOT
-      cd function-app && \
-      yarn install && \
-      func azure functionapp publish ${self.triggers.function_name}
-    EOT
-  }
+resource "aws_lambda_permission" "allow_cloudwatch" {
+  statement_id  = "AWSEvents_trigger-lambda"
+  action        = "lambda:InvokeFunction"
+  function_name = "${aws_lambda_function.lambda_function.function_name}"
+  principal     = "events.amazonaws.com"
+  source_arn    = "${aws_cloudwatch_event_rule.this.arn}"
 }
