+++ examples/simple/98-variables.tf
@@ -1,6 +1,6 @@
-variable "location" {
+variable "aws_region" {
   type    = string
-  default = "italynorth"
+  default = "eu-south-1"
 }
 
 variable "tags" {
@@ -15,5 +15,17 @@ variable "tags" {
 variable "prefix" {
   description = "Resorce prefix"
   type        = string
-  default     = "auditLogs-es-d"
+  default     = "adl"
 }
+
+variable "lambda" {
+  type = object({
+    role_name   = optional(string),
+    policy_name = optional(string)
+  })
+  default = {
+    role_name   =  "audit-logs-lambda-role",
+    policy_name = "audit-logs-lambda-policy"
+  }
+}
+
