--- examples/simple/99-outputs.tf
@@ -1,3 +1,0 @@
-output "resource_group_name" {
-  value = azurerm_resource_group.rg.name
-}
