--- examples/simple/backend.ini
@@ -1,1 +1,0 @@
-subscription=devopslab
