--- examples/simple/function-app/.node-version
@@ -1,1 +1,0 @@
-20.13.1
