--- examples/simple/function-app/.nvmrc
@@ -1,1 +1,0 @@
-20.13.1
