--- examples/simple/function-app/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "start": "func start",
-    "test": "echo \"No tests yet...\""
-  },
-  "dependencies": {
-    "@azure/functions": "^4.0.0",
-    "applicationinsights": "^2.1.6"
-  },
-  "devDependencies": {
-    "azure-functions-core-tools": "^4.x"
-  }
-}
